requirements.txt
python-telegram-bot
singleton_decorator
sqlalchemy
python-dotenv
logtail-python
loguru
openai
pytz
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-secret-manager
matplotlib
python-telegram-bot[job-queue]
flask[async]
aiohttp
flask
requests
alembic
driver
dataclasses_json
redis
pytest
pytest-asyncio
pytest-mock
pytest-cov
pytest-env
freezegun
faker